proper use of size_t

+++ trilogy-llvm/core/trilogy_array.h
@@ -7,20 +7,20 @@ trilogy_array_init(trilogy_value* tv, trilogy_array_value* arr);
 
 trilogy_array_value* trilogy_array_init_empty(trilogy_value* tv);
 
-trilogy_array_value* trilogy_array_init_cap(trilogy_value* tv, uint64_t cap);
+trilogy_array_value* trilogy_array_init_cap(trilogy_value* tv, size_t cap);
 
 trilogy_array_value*
 trilogy_array_clone_into(trilogy_value* tv, trilogy_array_value* arr);
 
-uint64_t trilogy_array_len(trilogy_array_value* tv);
-uint64_t trilogy_array_cap(trilogy_array_value* tv);
-uint64_t trilogy_array_resize(trilogy_array_value* tv, uint64_t cap);
-uint64_t trilogy_array_reserve(trilogy_array_value* tv, uint64_t cap);
+size_t trilogy_array_len(trilogy_array_value* tv);
+size_t trilogy_array_cap(trilogy_array_value* tv);
+size_t trilogy_array_resize(trilogy_array_value* tv, size_t cap);
+size_t trilogy_array_reserve(trilogy_array_value* tv, size_t cap);
 
 void trilogy_array_push(trilogy_array_value* arr, trilogy_value* tv);
 void trilogy_array_append(trilogy_array_value* arr, trilogy_value* tv);
 void trilogy_array_at(
-    trilogy_value* tv, trilogy_array_value* arr, uint64_t index
+    trilogy_value* tv, trilogy_array_value* arr, size_t index
 );
 
 int trilogy_array_compare(trilogy_array_value* lhs, trilogy_array_value* rhs);
