+++ trilogy-llvm/core/trilogy_string.c
@@ -31,7 +31,7 @@ trilogy_string_clone_into(trilogy_value* tv, const trilogy_string_value* orig) {
 }
 
 trilogy_string_value* trilogy_string_init_from_c(trilogy_value* tv, char* s) {
-    uint64_t len = (uint64_t)strlen(s);
+    size_t len = (size_t)strlen(s);
     trilogy_string_value* str = malloc_safe(sizeof(trilogy_string_value));
     str->len = len;
     str->contents = malloc_safe(sizeof(char) * len);
@@ -46,9 +46,9 @@ char* trilogy_string_as_c(trilogy_string_value* str) {
     return ptr;
 }
 
-uint64_t trilogy_string_len(trilogy_string_value* str) { return str->len; }
+size_t trilogy_string_len(trilogy_string_value* str) { return str->len; }
 
-uint32_t trilogy_string_at(trilogy_string_value* str, uint64_t index) {
+uint32_t trilogy_string_at(trilogy_string_value* str, size_t index) {
     assert(index < str->len);
     // TODO: properly support Unicode characters, instead of doing this on bytes
     return (uint32_t)str->contents[index];
@@ -57,7 +57,7 @@ uint32_t trilogy_string_at(trilogy_string_value* str, uint64_t index) {
 int trilogy_string_compare(
     trilogy_string_value* lhs, trilogy_string_value* rhs
 ) {
-    uint64_t len = lhs->len < rhs->len ? lhs->len : rhs->len;
+    size_t len = lhs->len < rhs->len ? lhs->len : rhs->len;
     int cmp = strncmp(lhs->contents, rhs->contents, len);
     if (cmp != 0) return cmp;
     if (lhs->len < rhs->len) return -1;
@@ -78,9 +78,9 @@ trilogy_string_value* trilogy_string_assume(trilogy_value* val) {
 trilogy_string_value* trilogy_string_concat(
     trilogy_value* rt, trilogy_string_value* lhs, trilogy_string_value* rhs
 ) {
-    uint64_t room = UINT64_MAX - lhs->len;
+    size_t room = SIZE_MAX - lhs->len;
     if (rhs->len > room) internal_panic("string length limit\n");
-    uint64_t len = lhs->len + rhs->len;
+    size_t len = lhs->len + rhs->len;
     char* ptr = malloc_safe(sizeof(char) * len);
     strncpy(ptr, lhs->contents, lhs->len);
     strncpy(ptr + lhs->len, rhs->contents, rhs->len);
